+++ src/main/java/com/kurly/api/jpa/repository/OptionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OptionRepository extends JpaRepository<Options ,Integer> {
+public interface OptionRepository extends JpaRepository<Options ,Long> {
 
 
 }
